+++ tests/urls.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 from django.urls import include, path
+from drf_spectacular.views import SpectacularAPIView
 from rest_framework.routers import DefaultRouter
 
 from tests.drf_app import PatientViewSet
@@ -7,4 +8,8 @@ from tests.drf_app import PatientViewSet
 router = DefaultRouter()
 router.register("patients", PatientViewSet, basename="patient")
 
-urlpatterns = [path("admin/", admin.site.urls), path("api/", include(router.urls))]
+urlpatterns = [
+    path("admin/", admin.site.urls),
+    path("api/schema/", SpectacularAPIView.as_view(), name="schema"),
+    path("api/", include(router.urls)),
+]